Move grid reader into file interface
