+++ fehm_toolkit/file_interface/__init__.py
@@ -1,3 +1,4 @@
 from .compact_node_data import write_compact_node_data
 from .fehm import read_fehm
+from .grid import read_grid
 from .zone import read_zones
